backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
async-timeout==5.0.1
bcrypt==3.2.2
billiard==4.2.4
celery==5.6.2
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cryptography==46.0.5
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.135.1
fastapi-cache2==0.2.2
flower==2.0.1
greenlet==3.3.2
h11==0.16.0
humanize==4.15.0
idna==3.11
joblib==1.5.3
kombu==5.6.2
lightgbm==4.6.0
numpy==2.2.6
packaging==26.0
pandas==2.3.3
passlib==1.7.4
pendulum==3.2.0
prometheus_client==0.24.1
prompt_toolkit==3.0.52
psycopg2==2.9.11
pyasn1==0.6.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pyotp==2.9.0
pypdf==6.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
pytz==2026.1.post1
redis==4.6.0
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
SQLAlchemy==2.0.48
starlette==0.52.1
threadpoolctl==3.6.0
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.41.0
vine==5.1.0
wcwidth==0.6.0
